admin-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
httpx==0.28.1
trafilatura==2.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
